+++ mc/src/filters.c
@@ -11,10 +11,9 @@
 
 
 /** \copydoc lpf_1st_init */
-void lpf_1st_init(Lpf1st_Obj_t* const lpf_1st_inst, double y, double u, 
-				double a0, double a1, double b1)
-
-				{
+void lpf_1st_init(Lpf1st_Obj_t* const lpf_1st_inst, float32_t y, float32_t u, 
+				float32_t a0, float32_t a1, float32_t b1)
+{
 	lpf_1st_inst->y = y;
 	lpf_1st_inst->u = u;
 
@@ -24,9 +23,9 @@ void lpf_1st_init(Lpf1st_Obj_t* const lpf_1st_inst, double y, double u,
 }
 
 /** \copydoc lpf_1st_update */
-void lpf_1st_update(Lpf1st_Obj_t* const lpf_1st_inst, double u)
+void lpf_1st_update(Lpf1st_Obj_t* const lpf_1st_inst, float32_t u)
 {
-	double u_k1 = lpf_1st_inst->u;
+	float32_t u_k1 = lpf_1st_inst->u;
 	lpf_1st_inst->y = lpf_1st_inst->b1 * lpf_1st_inst->y;
 	lpf_1st_inst->y += lpf_1st_inst->a0 * u;
 	lpf_1st_inst->y += lpf_1st_inst->a1 * u_k1;
